refactor(blog-post): 포스트 상세 화면에서 헤더 영역을 개별 컴포넌트로 분리 (#16)
포스트 상세 화면에서 역할과 책임에 따라
헤더 영역을 개별 컴포넌트로 분리하여
`blog-post.tsx`의 사이즈를 줄이고 가독성을 높임
